Update .travis.yml
Remove use of deprecated sudo line and specify build on default Ubuntu build environment.  Specify the use of an updated toolchain so that Travis CI can add gfortran 6.

In previous builds, getting gcc-7 from homebrew finished in minutes because it was a pre-built bottle.  The last run timed out because homebrew was compiling gcc-7.  The log mentioned that the XCode installation was out-of-date.  Updated to newer XCode, which implies a newer macOS and to gcc-8.